+++ src/vlrdevapi/_news/__init__.py
@@ -1,4 +1,5 @@
-from vlrdevapi._news.models import News, NewsPage
+from vlrdevapi._news.article.models import NewsArticle
+from vlrdevapi._news.list.models import News, NewsPage
 from vlrdevapi._news.namespace import NewsNamespace
 
-__all__ = ["News", "NewsNamespace", "NewsPage"]
+__all__ = ["News", "NewsArticle", "NewsNamespace", "NewsPage"]